--- a/src/main/scala/firrtl/transforms/Dedup.scala
+++ b/src/main/scala/firrtl/transforms/Dedup.scala
@@ -427,7 +427,11 @@ object DedupModules {
     val dedupedName2module = tag2name.map({ case (tag, name) => name -> DedupModules.dedupInstances(top, name, moduleMap, name2name, renameMap) })
 
     // Build map from original name to corresponding deduped module
-    val name2module = tag2all.flatMap({ case (tag, names) => names.map(n => n -> dedupedName2module(tag2name(tag))) })
+    // It is important to flatMap before looking up the DefModules so that they aren't hashed
+    val name2module: Map[String, DefModule] =
+      tag2all.flatMap { case (tag, names) => names.map(_ -> tag) }
+             .mapValues(tag => dedupedName2module(tag2name(tag)))
+             .toMap
 
     // Build renameMap
     val indexedTargets = mutable.HashMap[String, IndexedSeq[ReferenceTarget]]()
@@ -439,7 +443,7 @@ object DedupModules {
       }
     }
 
-    name2module.toMap
+    name2module
   }
 
   def computeIndexedNames(main: String, m: DefModule): IndexedSeq[ReferenceTarget] = {