@albert-magyar 2020-05-13
This should backport cleanly to 1.2.x